Fix dlogutil EPIPE handling.
EPIPE used to be wrongly considered an error.

In fact, it signals that the cyclic buffer has
made a full turn since the previous read, thus
overwriting the previous content.

Since dlogutil just dumps the contents of the
buffer and does not care about any particular
line being overwritten, it can safely be ignored.

Change-Id: I44cdb1e280957aace00c5a0cf4eadaa9a4a4133d
Signed-off-by: Michal Bloch <m.bloch@samsung.com>